+++ cmake/build_server_docker_everything.sh
@@ -4,10 +4,10 @@
 set -x
 
 # Fletch, VIAME, CMAKE system deps
-./viame/cmake/linux_install_deps_ubuntu.sh
+./viame/cmake/build_server_ubuntu_deps.sh
 
 # Install CMAKE
-./viame/cmake/linux_build_and_install_cmake.sh
+./viame/cmake/build_server_linux_cmake.sh
 
 # Update VIAME sub git deps
 cd /viame/
@@ -15,12 +15,9 @@ git submodule update --init --recursive
 mkdir build
 cd build
 
-# Configure Paths [should be removed when no longer necessary by fletch]
+# Add VIAME and CUDA paths to build
 export PATH=$PATH:/usr/local/cuda/bin:/viame/build/install/bin
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/cuda/lib64
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/viame/build/install/lib:/viame/build/install/lib/python3.6
-export C_INCLUDE_PATH=$C_INCLUDE_PATH:/viame/build/install/include/python3.6m
-export CPLUS_INCLUDE_PATH=$CPLUS_INCLUDE_PATH:/viame/build/install/include/python3.6m
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/viame/build/install/lib:/usr/local/cuda/lib64
 
 # Configure VIAME
 cmake ../ -DCMAKE_BUILD_TYPE:STRING=Release \
@@ -28,11 +25,10 @@ cmake ../ -DCMAKE_BUILD_TYPE:STRING=Release \
 -DVIAME_FIXUP_BUNDLE:BOOL=OFF \
 -DVIAME_VERSION_RELEASE:BOOL=ON \
 -DVIAME_ENABLE_BURNOUT:BOOL=ON \
--DVIAME_ENABLE_CAFFE:BOOL=OFF \
 -DVIAME_ENABLE_CUDA:BOOL=ON \
 -DVIAME_ENABLE_CUDNN:BOOL=ON \
--DVIAME_ENABLE_DIVE:BOOL=OFF \
 -DVIAME_ENABLE_DARKNET:BOOL=ON \
+-DVIAME_ENABLE_DIVE:BOOL=OFF \
 -DVIAME_ENABLE_DOCS:BOOL=OFF \
 -DVIAME_ENABLE_FFMPEG:BOOL=ON \
 -DVIAME_ENABLE_FFMPEG-X264:BOOL=ON \
@@ -46,12 +42,11 @@ cmake ../ -DCMAKE_BUILD_TYPE:STRING=Release \
 -DVIAME_ENABLE_OPENCV:BOOL=ON \
 -DVIAME_OPENCV_VERSION:STRING=3.4.0 \
 -DVIAME_ENABLE_PYTHON:BOOL=ON \
--DVIAME_ENABLE_PYTHON-INTERNAL:BOOL=ON \
--DVIAME_PYTHON_VERSION:STRING=3.6.15 \
+-DVIAME_PYTHON_BUILD_FROM_SOURCE:BOOL=OFF \
 -DVIAME_ENABLE_PYTORCH:BOOL=ON \
--DVIAME_ENABLE_PYTORCH-INTERNAL:BOOL=ON \
--DVIAME_PYTORCH_VERSION:STRING=1.10.2 \
--DVIAME_ENABLE_PYTORCH-DISABLE-NINJA=ON \
+-DVIAME_PYTORCH_BUILD_FROM_SOURCE:BOOL=ON \
+-DVIAME_PYTORCH_VERSION:STRING=2.5.1 \
+-DVIAME_PYTORCH_DISABLE_NINJA=ON \
 -DVIAME_ENABLE_PYTORCH-MMDET:BOOL=ON \
 -DVIAME_ENABLE_PYTORCH-NETHARN:BOOL=ON \
 -DVIAME_ENABLE_PYTORCH-PYSOT:BOOL=ON \